Added graph drawing code.  See description
The number of bars is hardcoded in GraphView.swift.  Make sure the number of labels and number of values are the same.  In MCQGraphVC.swift there is a method which modifies the widths of the bars (from 0 to 1).  You can do this from any code.  I think you only need graphView.setNeedsDisplay() in MCQGraphVC.swift because in other views it is not visible, so will need to be redrawn anyway.